+++ src/main/java/com/supermartijn642/oregrowth/OreGrowth.java
@@ -52,6 +52,7 @@ public class OreGrowth implements ModInitializer {
 
     public static void registerGenerators(){
         GeneratorRegistrationHandler handler = GeneratorRegistrationHandler.get(MODID);
+        handler.addGenerator(OreGrowthAtlasSourceGenerator::new);
         handler.addGenerator(OreGrowthBlockStateGenerator::new);
         handler.addGenerator(OreGrowthItemInfoGenerator::new);
         handler.addGenerator(OreGrowthLanguageGenerator::new);