+++ changelog.md
@@ -1,2 +1,2 @@
 ### Ore Growth 1.1.5
-- Initial release of Ore Growth for Minecraft 1.21.5
+- Initial release of Ore Growth for Minecraft 1.21.6 & 1.21.7